dom/indexedDB/test/test_cursors.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const name = window.location.pathname;
      const description = "My Test Database";
      const keys = [1, -1, 0, 10, 2000, "q", "z", "two", "b", "a"];
      const sortedKeys = [-1, 0, 1, 10, 2000, "a", "b", "q", "two", "z"];

      is(keys.length, sortedKeys.length, "Good key setup");

      let request = indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.result;

      request = db.createObjectStore("foo", "");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let objectStore = event.result;

      let keyIndex = 0;

      for (let i in keys) {
        request = objectStore.add("foo", keys[i]);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          if (++keyIndex == keys.length) {
            testGenerator.next();
          }
        };
      }
      yield;

      keyIndex = 0;

      request = objectStore.openCursor();
      request.onerror = errorHandler;
      request.onsuccess = function (event) {
        let cursor = event.result;
        if (cursor) {
          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          let ok = cursor.continue();
          is(ok, true, "Correct return from continue");

          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          keyIndex++;
        }
        else {
          testGenerator.next();
        }
      }
      yield;

      is(keyIndex, keys.length, "Saw all added items");

      keyIndex = 4;

      let range = indexedDB.makeBoundKeyRange(2000, "q");
      request = objectStore.openCursor(range);
      request.onerror = errorHandler;
      request.onsuccess = function (event) {
        let cursor = event.result;
        if (cursor) {
          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          let ok = cursor.continue();
          is(ok, true, "Correct return from continue");

          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          keyIndex++;
        }
        else {
          testGenerator.next();
        }
      }
      yield;

      is(keyIndex, 8, "Saw all the expected keys");

      keyIndex = 0;

      request = objectStore.openCursor();
      request.onerror = errorHandler;
      request.onsuccess = function (event) {
        let cursor = event.result;
        if (cursor) {
          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          let ok = keyIndex ? cursor.continue() : cursor.continue("b");
          is(ok, true, "Correct return from continue");

          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          keyIndex += keyIndex ? 1: 6;
        }
        else {
          testGenerator.next();
        }
      }
      yield;

      is(keyIndex, keys.length, "Saw all the expected keys");

      keyIndex = 0;

      request = objectStore.openCursor();
      request.onerror = errorHandler;
      request.onsuccess = function (event) {
        let cursor = event.result;
        if (cursor) {
          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          let ok = keyIndex ? cursor.continue() : cursor.continue(10);
          is(ok, true, "Correct return from continue");

          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          keyIndex += keyIndex ? 1: 3;
        }
        else {
          testGenerator.next();
        }
      }
      yield;

      is(keyIndex, keys.length, "Saw all the expected keys");

      keyIndex = 0;

      request = objectStore.openCursor();
      request.onerror = errorHandler;
      request.onsuccess = function (event) {
        let cursor = event.result;
        if (cursor) {
          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          let ok = keyIndex ? cursor.continue() : cursor.continue("c");
          is(ok, true, "Correct return from continue");

          is(cursor.key, sortedKeys[keyIndex], "Correct key");
          is(cursor.value, "foo", "Correct value");

          keyIndex++;
        }
        else {
          testGenerator.next();
        }
      }
      yield;

      is(keyIndex, keys.length, "Saw all the expected keys");

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
